requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Embeddings
google-generativeai==0.3.0
langchain==0.1.20
langchain-google-genai==0.1.0
langchain-core==0.1.30
langchain-text-splitters

# Graph & Orchestration
langgraph==0.0.60

# Postgres + pgvector
asyncpg==0.29.0
pgvector==0.2.5

# Valkey / Redis
redis==5.0.1

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
aiohttp==3.9.1
httpx==0.25.2
python-multipart==0.0.6

# Utilities
numpy==1.24.3
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
